+++ src/views/List/StockItem/SummaryInfo/SummaryContent.tsx
@@ -1,9 +1,10 @@
 import { useSelector } from 'react-redux';
 
 import { TableCell } from '../../../../components/Table';
-import { Input, TransformedValue } from '../../../../components/Input';
-import { NumberCell } from '../components';
+import { Input } from '../../../../components/Input/Input';
+import { TransformedValue } from '../../../../components/Input/BaseInput';
 import { selectStockInfoById } from '../../../../features/stockList/selectors';
+import { NumberCell } from '../components';
 import { useGetStockSummaryData } from './useGetStockSummaryData';
 
 type Props = {
